Added authentication, cover image uploads and restrictions.

+++ app/Http/Controllers/PostsController.php
@@ -7,6 +7,17 @@ use App\Post;
 
 class PostsController extends Controller
 {
+
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -50,13 +61,32 @@ class PostsController extends Controller
         $this->validate($request, [
             'title' => 'required',
             'body' => 'required',
+            'cover_image' => 'image|nullable|max:5999',
         ]);
 
+        
+        // Handle file upload
+        if($request->hasFile('cover_image')) {
+            // Get filename with ext
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // Get the filename only
+            $filename = pathInfo($filenameWithExt, PATHINFO_FILENAME);
+            // Get the ext only
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            // Get file name to store
+            $fileNameToStore = $filename . '_' . time() . '.' . $extension;
+            // Upload image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        } else {
+            $fileNameToStore = 'noimage.jpg';
+        }
+        
         // Create Post entry
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
         $post->user_id = auth()->user()->id;
+        $post->cover_image = $fileNameToStore;
         $post->save();
 
         return redirect('/posts')->with('success', 'Post Created');
@@ -87,6 +117,11 @@ class PostsController extends Controller
         // Find the post by id
         $post = Post::find($id);
 
+        // Check for correct user
+        if(auth()->user()->id !=$post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
+
         return view('posts.edit')->with('post', $post);
     }
 
@@ -103,12 +138,34 @@ class PostsController extends Controller
         $this->validate($request, [
             'title' => 'required',
             'body' => 'required',
+            'cover_image' => 'image|nullable|max:5999',
         ]);
 
+        
+        // Handle file upload
+        if($request->hasFile('cover_image')) {
+            // Get filename with ext
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // Get the filename only
+            $filename = pathInfo($filenameWithExt, PATHINFO_FILENAME);
+            // Get the ext only
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            // Get file name to store
+            $fileNameToStore = $filename . '_' . time() . '.' . $extension;
+            // Upload image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        } else {
+            $fileNameToStore = 'noimage.jpg';
+        }
+        
         // Create Post entry
         $post = Post::find($id);
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        $post->user_id = auth()->user()->id;
+        // Check if the cover_image has been updated
+        if($fileNameToStore != 'noimage.jpg')
+            $post->cover_image = $fileNameToStore;
         $post->save();
 
         return redirect('/posts')->with('success', 'Post Updated');
@@ -123,6 +180,12 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        // Check for correct user
+        if(auth()->user()->id !=$post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
+
         $post->delete();
 
         return redirect('/posts')->with('success', 'Post Deleted');
